YAY-21 refactor: delete ProfileHeaderContainer
