+++ src/widgets/Profile/index.ts
@@ -1,2 +1,1 @@
 export { ProfileView } from './ui/ProfileView'
-export { ProfileHeaderContainer } from './ui/ProfileHeaderContainer'
